requirements.txt
ollama
langchain
langchain-huggingface
langchain-community
sentence-transformers
faiss-cpu
Pillow
pytesseract
unstructured
